Move out table options from util to stoarge/constants
