+++ Cargo.toml
@@ -47,6 +47,7 @@ members = [
     "src/query/legacy-planners",
     "src/query/settings",
     "src/query/storages/context",
+    "src/query/storages/constants",
     "src/query/storages/fuse",
     "src/query/storages/fuse-meta",
     "src/query/storages/hive",
