+++ src/query/storages/constants/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "common-storages-constants"
+version = { workspace = true }
+authors = { workspace = true }
+license = { workspace = true }
+publish = { workspace = true }
+edition = { workspace = true }
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+[lib]
+doctest = false
+test = false
+
+[dependencies]
+
+once_cell = "1.15.0"
