+++ src/query/storages/fuse/src/lib.rs
@@ -31,12 +31,11 @@ use common_catalog::table::Table;
 use common_catalog::table::TableStatistics;
 pub use common_catalog::table_context::TableContext;
 use common_catalog::table_mutator::TableMutator;
-use common_storages_util::table_option_keys;
+pub use common_storages_constants::*;
 pub use constants::*;
 pub use fuse_part::ColumnLeaf;
 pub use fuse_part::ColumnLeaves;
 pub use fuse_table::FuseTable;
-pub use table_option_keys::*;
 
 mod sessions {
     pub use common_catalog::table_context::TableContext;
