+++ src/query/storages/share/Cargo.toml
@@ -14,6 +14,7 @@ test = false
 [dependencies]
 common-exception = { path = "../../../common/exception" }
 common-meta-app = { path = "../../../meta/app" }
+common-storages-constants = { path = "../constants" }
 common-storages-util = { path = "../util" }
 
 opendal = { version = "0.19", features = ["layers-retry"] }
