+++ src/query/storages/util/Cargo.toml
@@ -19,5 +19,4 @@ common-exception = { path = "../../../common/exception" }
 common-fuse-meta = { path = "../fuse-meta" }
 
 async-trait = { version = "0.1.57", package = "async-trait-fn" }
-once_cell = "1.15.0"
 opendal = { version = "0.19", features = ["layers-retry"] }
