+++ src/query/storages/util/src/lib.rs
@@ -14,5 +14,3 @@
 
 pub mod cached_reader;
 pub mod file_meta_data_reader;
-pub mod table_option_keys;
-pub mod table_storage_prefix;
